:bug: [#5735] Fix crash in conversion of values to Python domain due to non-existing variables
This is an edge case for steps - containing components with variable actions - that were already submitted, but later marked
as not applicable.

The "step not applicable" action deletes the related submission value variables by setting an empty `FormioData` instance to
the step data. This also removes the variables from the submission state, so we skip the conversion for non-existing variables.
Question: is it worth considering to not remove the variables completely from the state, but just set their primary keys to
`None`? :thinking_face:

Note that this is just a bandaid fix, and ideally we would like to skip any logic triggers and actions that involve variables
of components on non-applicable steps. We need a dependency graph to implement this properly, as logic rules and actions might
not be limited to the non-applicable step.