:label: Ban typing.cast from the codebase
typing.cast always overrules the type checker and allows you to write
lies way too easily. The code being cast may also change, and the
CI/type checker will never report if the cast is no longer necessary
or changed in nature making it more or less wrong. Additionally, it's
hard to keep track of why the cast is/was necessary in this context.

Instead, it's usually preferred to narrow types for the type checker
using asserts, which are stripped out in production so the perf
hit is irrelevant. If that is not possible, using a specific type
ignore comment is better because it at least points at the kind of
problem is being resolved, and we can (in the future) check if the
type ignores go down rather than up when the majority of the codebase
is checked in CI *and* tooling in the ecosystem improves.

The rule suppressions in the registrations app are necessary because
TypedDicts aren't easy/pleasant to work with, especially when
correlating classes with behaviour and value types with them. The
types in the openforms.formio package are far from ideal, but this
is too much to address now.