Fixed date format and missing keys in family members
StUF-BG has a different format for dates, it uses yyyymmdd, while
HaalCentraal uses yyyy-MM-dd. This was not handled properly as it used
only the yyyy-MM-dd format.

StUF-BG does not have dateOfBirthPrecision so this was missed in some
registration backends and we were trying to access it (KeyError).

(cherry picked from commit 505a0685b11e70875b135cc5adaf9492215c615a)