[#5748] Omit deceased property in partners and children before validation
In HaalCentraal we have to make an extra call to retrieve the deceased children. This is not happening with StUF-BG where we have this information in the XML response and that's why we have an extra property (deceased) in the immutable variable, which always takes a default value (None) if not provided.
This makes our validation check for tampered data to fail. Have to remove this key before we do the check between the initial value and the mutated one.

Backport-of: #5766