:bug: [#5765] Fix submission data for radio/select/selectboxes components with dynamic options not passing step validation
The fix for CVE-2025-64515 introduced an extra logic evaluation in the submission step serializer
(`SubmissionStepSerializer.to_internal_value`) to prevent tampering with prefill data. Because we
do not pass the (unvalidated) step data there, the configuration is not representative for the one
upon step validation.

To ensure that we are not working with this non-representative configuration, we need to force a
re-evaluation of form logic before running the formio validation. Note that this will have an
impact on performance, but it is only during step validation and not during the check logic calls,
so it will not be as impactful.

Backport-of: #5768