:recycle: [#2409] Ensure user-defined variables are persisted upon step submission
Previously this was only done upon completing the entire submission. This change will greatly
simplify getting logic rules per step, as we can rely on user-defined variables already being
resolved after each step. This means we do not have to perform tricky graph analysis to determine
which rules are dependent on user-defined variables, and always execute those.

It seems the check on the last form step in `persist_user_defined_variables` is very outdated
and no longer relevant, so I have removed it