:triangular_flag_on_post: [#2409] Add feature flag for persisting user-defined variables upon step completion
We want this to be the new behaviour, so we use opt-in by default. If it turns out to be completely broken,
municipalities have an easy escape hatch.

Added a call to `check_submission_logic` before persisting variables in the submission completion mixin, as this
was the previous behaviour of `persist_user_defined_variables`.