:recycle: [#5747] Refactor bulk updating submission value variables from data
The previous implementation had the (unwanted?) side effect of persisting variables to the database if their keys were not in the provided `data`.
This created issues during persisting of user-defined variables, as variables related to hidden component or not-applicable steps would be saved
to the database unexpectedly.

This change ensures we do not create/update variables that were not present in `data`.

Also refactored handling of deleting variables. We still delete existing variables if their key is not present in `data` (and the
`reset_missing_variables` flag is set to `True`), but we do not delete the variable from the state anymore. Instead, we set the primary key to
`None` and reset the value to the initial value. This is to ensure we always have a submission value variable instance available for all form
variable.