:recycle: [#5747] Refactor bulk updating submission value variables from data
The previous implementation had the (unwanted) side effect of persisting variables to the database if their keys were not in the provided `data`.
This created issues during persisting of user-defined variables, as variables related to hidden component or not-applicable steps would be saved
to the database unexpectedly.

Now, we only delete existing variables if their key is not present in `data` (and the `delete_missing_variables` flag is set). All entries in
`data` for which we do not have a variable will be ignored. We can do this because the variables are fetched from the state, which gives a
complete picture of all relevant variables (either from the database or in-memory).

Also removed setting the value to the initial value of the form variable, as this was not persisted to the database anyway. I'm assuming this
was originally implemented to reset value, but without comments, a docstring, and a commit message, one can only guess :shrug: