:adhesive_bandage: [#5803] Ensure context for Objects API v1 templates includes only JSON values for submission data
Python date/time/datetime objects (which are now used inside templates because of
the rework done in GH-2324) are automatically formatted according to the locale if
there is no format specified in the template variables.

Unfortunately, this causes issues for existing templates that rely on the format being
an ISO-8601 string.

The proper solution to this would be to adjust all templates (either manually or
automatically with a script) by adding a `|date:'Y-m-d'` to the date template
variables, but dealing with date-related objects which are part of editgrid fields
seems a rather tricky thing to do inside the templates themselves. Therefore, we
apply this bandaid fix to revert back to the behaviour from before 3.3.0

I went for this custom override of `get_variables_for_context` instead of adding an extra
argument `to_json` to the original, because this is a bandaid fix. We would still like to
always use date-related objects in the context for templates.

Backport-of: #5804