[#5776] Added user defined variables to the stuf-zds extraelementen dropdown
The purpose of this is to give the opportunity to the users to change
the names of the user defined variables that they choose to be in the template and
specifically in the extraElementen. We also removed the filter in the
backend, which was accepting only the plugin payment variables in the
extraElementen (even if the user had selected somthing else).

Now we have the default payment variables (for the plugin) with a
default stuf name and the user can add other variables mappings too.