:white_check_mark: [#5729] Added tests for profile component/CommunicationPreferences pre-registration hook usage while authenticated using kvk.
Renamed the existing tests to make it explicit which authentication attribute was used.

There are only new kvk tests added for situations that are different.
Whether you login using bsn or kvk, we process already existing digital addresses the same way, so adding new tests with minimal difference (login with kvk instead of bsn) seems unnecessary.

Updated the vcr cassettes due to the renaming of tests, and new kvk specific tests. The only truely new cassettes are the cassettes for "kvk" tests, the others are just renames/updates of the previous cassettes.