:art: Explicitly mark tasks to ignore results
The cleanup task takes care of explicitly forgetting them, but for all
the other tasks you typically want to ignore the result if it's not
accessed anywhere.

Most of our tasks are just 'fire and forget', and due to Redis keys
having a TTL it won't mess up memory usage of Redis, but it's good
practice to be aware of these semantics.